Arreglo Cedula.

+++ Conservatorio.DATOS/Persona.cs
@@ -6,7 +6,7 @@ namespace Conservatorio.DATOS
     {
         public virtual int IdPersona { get; set; }
         public virtual string Nombre { get; set; }
-        public virtual int? Cedula { get; set; }
+        public virtual long? Cedula { get; set; }
         public virtual bool Estado { get; set; }
         public virtual string Email { get; set; }
         public virtual string Direccion { get; set; }
